+++ Editor/ProjectAuditor.cs
@@ -162,6 +162,12 @@ namespace Unity.ProjectAuditor.Editor
                 {
                     onIncomingIssues = issues =>
                     {
+                        // update diagnostic priorities based on dependencies
+                        foreach (var issue in issues.Where(i => i.descriptor != null && !i.isCritical && i.dependencies != null && i.dependencies.IsPerfCritical()))
+                        {
+                            issue.isCritical = true;
+                        }
+
                         report.AddIssues(issues);
                         projectAuditorParams.onIncomingIssues?.Invoke(issues);
                     },